Added exception handling for Categories API endpoint

+++ src/main/java/com/boukriinfo/ecommerce/services/CategoryServiceImpl.java
@@ -1,6 +1,7 @@
 package com.boukriinfo.ecommerce.services;
 
 import com.boukriinfo.ecommerce.entities.Category;
+import com.boukriinfo.ecommerce.exceptions.CategoryNotFoundException;
 import com.boukriinfo.ecommerce.repositories.CategoryRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,6 +18,7 @@ import java.util.List;
 @Slf4j
 public class CategoryServiceImpl implements CategoryService {
     private CategoryRepository categoryRepository;
+
     @Override
     public Category saveCategory(Category category) {
         return categoryRepository.save(category);
@@ -29,14 +31,14 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category getCategory(Long categoryId) {
-        Category category = categoryRepository.findById(categoryId).orElseThrow(() -> new RuntimeException("Category by Id :"+categoryId+" not found"));
+        Category category = categoryRepository.findById(categoryId).orElseThrow(() -> new RuntimeException("Category by Id :" + categoryId + " not found"));
         return category;
     }
 
     @Override
     public Category updateCategory(Category category) {
-        Category categoryUpdate = categoryRepository.findById(category.getId()).orElseThrow(() -> new RuntimeException("Category by Id :"+category.getId()+" not found"));
-        if (categoryUpdate != null){
+        Category categoryUpdate = categoryRepository.findById(category.getId()).orElseThrow(() -> new CategoryNotFoundException("Category by Id :" + category.getId() + " not found"));
+        if (categoryUpdate != null) {
             categoryUpdate.setName(category.getName());
             categoryUpdate.setSlug(category.getSlug());
             categoryUpdate.setDescription(category.getDescription());
@@ -52,28 +54,44 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category updateDeletedCategory(Category category) {
-        category.setDeleted(true);
-        return categoryRepository.save(category);
+        Category categoryUpdate = categoryRepository.findById(category.getId()).orElseThrow(() -> new CategoryNotFoundException("Category by Id :" + category.getId() + " not found"));
+        if (categoryUpdate != null) {
+            categoryUpdate.setDeleted(true);
+            categoryRepository.save(categoryUpdate);
+            return categoryUpdate;
+        }
+        return null;
     }
 
     @Override
-    public void deleteCategory(Long categoryId) {
-        categoryRepository.deleteById(categoryId);
-
+    public boolean deleteCategory(Long categoryId) {
+        Category category = categoryRepository.findById(categoryId).orElseThrow(() -> new CategoryNotFoundException("Category by Id :" + categoryId + " not found"));
+        if (category != null) {
+            categoryRepository.deleteById(categoryId);
+            return true;
+        } else {
+            return false;
+        }
     }
+
 
     @Override
     public Page<Category> getAllCategoriesWithPage(int page, int size) {
-        return categoryRepository.findAllNotDeleted(PageRequest.of(page,size));
+        return categoryRepository.findAllNotDeleted(PageRequest.of(page, size));
     }
 
     @Override
     public Page<Category> getAllCategoriesWithSlugAndPage(String slug, int page, int size) {
-        return categoryRepository.findAllBySlugContainingAndNotDeleted(slug,PageRequest.of(page,size));
+        return categoryRepository.findAllBySlugContainingAndNotDeleted(slug, PageRequest.of(page, size));
     }
 
     @Override
     public List<Category> findAllCategoriesNotDeleted() {
-        return categoryRepository.findAllProductsNotDeleted();
+        List<Category> categories = categoryRepository.findAllProductsNotDeleted();
+        if (categories.isEmpty()) {
+            throw new CategoryNotFoundException("Aucune catégorie non supprimée n'a été trouvée.");
+        }
+        return categories;
     }
+
 }
